+++ tests/corpus/py_src_layout/src/pkg/__init__.py
@@ -1,0 +1,3 @@
+"""Package root of a src-layout project."""
+
+VERSION = "0.0.0"